Commenting out git add/commit calls when initalizing new repo.
The working folder is empty, so the add/commit is ignored by git.
If it is truly necessary to have an initial commit done within git
prior to processing the first AccuRev transaction, then some kind of
file will have to be added to the working folder prior to running
the add/commit calls. Perhaps an empty .gitignore file, though it
really should be empty as we don't want to skip anything that was
promoted to AccuRev while processing those transactions.